ssl_client: fix unsigned comparison

+++ libraries/SSLClient/src/ssl_client.cpp
@@ -226,7 +226,7 @@ int start_ssl_client(sslclient_context *ssl_client, const char *host, uint32_t p
         }
         unsigned char psk[MBEDTLS_PSK_MAX_LEN];
         size_t psk_len = strlen(psKey)/2;
-        for (int j=0; j<strlen(psKey); j+= 2) {
+        for (size_t j=0; j<strlen(psKey); j+= 2) {
             char c = psKey[j];
             if (c >= '0' && c <= '9') c -= '0';
             else if (c >= 'A' && c <= 'F') c -= 'A' - 10;
